refactor: convert interfaces to types and remove I prefix
- Removed 'I' prefix from type names following TypeScript best practices
- Converted interfaces to type aliases where appropriate
- Updated type exports to use explicit named exports
- Updated type documentation to reflect changes
- Maintained consistent type naming across the codebase
