+++ src/core/adapters/akamai/AkamaiKVStore.ts
@@ -1,4 +1,4 @@
-import { IKVStore } from '../../../types/cdn';
+import { KVStore } from '../../../types/cdn';
 import { Logger } from '../../../utils/logging/Logger';
 
 /**
@@ -19,9 +19,9 @@ interface AkamaiEnv {
 
 /**
  * AkamaiKVStore provides a concrete implementation for interacting with Akamai's KV store.
- * It implements the IKVStore interface to ensure consistent KV store operations across different platforms.
+ * It implements the KVStore interface to ensure consistent KV store operations across different platforms.
  */
-export class AkamaiKVStore implements IKVStore {
+export class AkamaiKVStore implements KVStore {
 	private readonly NOT_IMPLEMENTED = 'AkamaiKVStore is not implemented yet';
 	private logger?: Logger;
