+++ src/types/cdn/settings.ts
@@ -1,0 +1,13 @@
+/**
+ * Type for CDN variation settings
+ */
+type CDNVariationSettings = {
+  cdnExperimentURL?: string;
+  cdnResponseURL?: string;
+  cacheKey?: string;
+  forwardRequestToOrigin?: boolean;
+  cacheRequestToOrigin?: boolean;
+  isControlVariation?: boolean;
+}
+
+export type { CDNVariationSettings };
