+++ src/types/cdn/store.ts
@@ -1,0 +1,10 @@
+/**
+ * Type for key-value store operations
+ */
+type KVStore = {
+  get(key: string): Promise<string | null>;
+  put(key: string, value: string): Promise<void>;
+  delete(key: string): Promise<void>;
+}
+
+export type { KVStore };
