+++ src/types/index.ts
@@ -1,6 +1,6 @@
-export * from './cdn';
-export * from './core';
+export * from './api/index';
+export * from './cdn/index';
+export * from './core/index';
+export * from './events/index';
+export * from './http/index';
 export * from './optimizely';
-export * from './request';
-export * from './response';
-export * from './events';
